added LGPL v3 license
Since this piece of software touches a highly competitive market (semiconductor
industry), I feel obliged to protect the code from any legal shenanigans, patent
trolling, or similar issues. The LGPL also ensures that the code may be used in
commercial products, but that improvements to the code made by the vendor must
be returned to the community. A fair deal.