VHDL lexer is improving
Things are fairly fleshed out as of now. No error reporting is implemented yet, but
it should be possible to lex every VHDL source file out there. Next up is tagging
identifiers which are reserved keywords with a special token type, such that the
parser may operate on them as quickly as possible.