bypassing KeywordMapper, seems broken
The llhd/vhdl/test/parser.cpp freezes at seemingly random points in time. This
might be an issue with std::string(bc,c), where the process seems to hang on a
regular basis. However it might also be possible that the KeywordMapper causes
the freeze with its search algorithm. Disabled the KeywordMapper for the time
being to get things to test properly.