added documentation to utf16
The functions decodePedantic(), decode() and encode() lacked documentation.
This is now fixed. Also added a TODO file that captures some of the current
development leads. This is a temporary solution, in the long run GitHub issues
are a better way to track open leads of development, especially when multiple
programmers are involved.