implemented keyword table lookup
Added vhdl/tblgen-keywords.cpp that generates a lookup function for keywords.
Its structure is fairly efficient, performing a branch for every character of
the input iterator. This generates the vhdl/keywords.hpp file which needs to be
in the source tree as a header file, since the mapping function is a template.

The LexerNovum uses this function to translate identifiers into keywords as soon
as they are read. To do this, the start and end pointers into the source buffer
are wrapped in unicode::casefold_iterator objects. These are then passed to the
lookupKeyword() function.

Also moved the TokenType enum into its own file vhdl/TokenType.hpp.