fixed src mgr locations
Fixed the calculation of locations in the SourceManager. Also added the
PresumedRange class that represents a range of PresumedLocation objects. Added
operator<< for the classes in SourceLocation.hpp for convenience. Also added a
unit test that checks the integrity of the locations the SourceManager produces.
Helped debugging quite a lot.