first simulation output
The llhd-sim2 program now outputs a clock signal generated through the
simulation event queue. The next step would be the implementation of
instructions and assignment of their results to signals. This would allow for
pretty fancy and complex logic circuits to be simulated already. Maybe the
delay instruction would be the easiest to implement first?