first steps towards registers and storage
Cleaned up a lot of the code. Introduced the edge, rise, and fall instructions
that extract the corresponding edges from a signal, e.g. a clock. Also added
the st instruction that moves data into a register ('reg' signal) while a given
trigger signal is high. This trigger may be the result of one of the edge
detection instructions, or a more complicated signal combination that also
allows asynchronous resets.

The sim2.cpp program now creates a bunch of signals, some of which describe a 1
bit counter with asynchronous reset. There is significant overhead in the
number of signals required to achieve asynchronous resets, the overhead goes
away as the number of registers grows, since the generation of the trigger
signal is only required once. Nevertheless, the syntax could me more concise.
This approach also does not clearly designate a signal as a clock. The trigger
tends to be the result of a boolean gate, which makes it difficult to explicitly
describe a gated clock. Maybe the edge detection should be merged with the
st command directly.