added grammar file
Summarized the basics of the LLHD assembly grammar into a separate text file.
Seems like the syntax may be kept pretty straightforward, especially if we
limit the language strictly to hardware description. Excluding testing
functionality makes the language very compact and easy to process. Extensive
testing would still be achievable through the use of a scripting language such
as Lua that features a more comfortable working environment. VHDL and Verilog
compilers could then produce both a .llhd and a .lua file which capture the
whole of the code.