Add virtual machine experimental code
Create a directory for experimental programs and code doodles. Add a
virtual machine that allows basic arithmetic, branching, and
comparisons. The VM contains only basic types at the moment, but shall
be extended in the future to provide access to logic and logic vector
types. Eventually, all hardware simulation can be translated into code
that runs on the virtual machine.

* Add exp/vm.cpp
* Add exp/CMakeLists.txt to compile experimental code
* Modify CMakeLists.txt to build programs in exp