Add assertion header
Define a few macros that perform static and runtime assertions. Call
them llhd[_static]_assert[_msg], such that throughout the codebase, we
only have to refer to these macros. Note that on compilers other than
clang, the calls to static_assert probably have to be implemented as
runtime assertions. Need an #if/#endif for that.

* Add src/utils/assert.hpp