Add build versioning
Modify the CMakeLists.txt file to keep track of the major, minor, and
patch version of LLHD, as well as a version suffix and the entire
version string. Load the git branch and commit at the time of the build
into the LLHD_GIT_COMMIT and LLHD_GIT_BRANCH variables. Add an option
that enables *nightly* builds. If the option is set, the version string
is extended with a timestamp and abbreviated git commit for clarity.

* Add variables and calls to git to CMakeLists.txt