Make legacy code base and tests optional
Split the build of the source code and the tests into two separate
blocks that can be enabled with the options LLHD_BUILD_LEGACY and
LLHD_BUILD_TESTS. The source files in the llhd/ shall be migrated into a
new source tree step by step. Also make the documentation optional via
the LLHD_BUILD_DOC option.

Wrap the tests of the legacy codebase in an if()/endif() block such that
they are only built when the legacy codebase is built alongside of them.
Define a macro that allows for the easy definition of test cases and add
a dummy test as a proof that it works.

* Add options and if blocks to CMakeLists.txt
* Adjust test/CMakeLists.txt