Use lcov for coverage analysis
Switch to lcov for coverage analysis and rewrite parts of the CMakeLists
file such that it defines additional targets for analysis setup,
capture, and report generation. Adjust the Travis CI script to take
advantage of the new setup. Also switch to the new Travis CI
infrastructure that runs on containers.

* Add new targets to CMakeLists.txt
* Adjust .travis.yml accordingly