Properly zero coverage counters and reduce noise
The coverage target coverage-init does not properly zero the execution
counters, but rather just captures the initial baseline coverage state.
Adjust the target to execute lcov with the --zerocounters option, which
essentially removes all existing execution counts and thus also the
coverage data.

Pipe the output of lcov and genhtml into a log file in the build
directory, such that there is less noise in the console.

* Adjust coverage commands in CMakeLists.txt