Implement file reading utility
Add a function llhd::readfile that reads a file into an std::string or
an std::vector<char>. This is useful for parsing files, where the
entirety of the file is operated on anyway. Additionally, it should
prove easy to open the file through mmap in the future, thus speeding up
the reading time and making things more efficient.

* Add src/utils/readfile.{hpp,cpp}