Implement an assembly parser
Implement the grammar outlined in ASSEMBLY.md as a simple recursive
descent parser. The implementation is far from optimal, but makes heavy
use of the polymorphic types with value semantics described in the
previous commit, which makes the code easy to read. Helper functions
could heavily improve and shorten the parser code, especially the
diagnostics parts.