[rebase] working on desequentialization
The simplify() function currently creates an infinite recursion. It's
probably time to write a dedicated, small boolean expression solver that
can be used to symbolically evaluate a graph of functions and supports
AND and OR operations with more than two arguments. It should also be
easier to perform various simplifications on such a representation,
rather than having to walk an entire tree of binary operations. Call it
SBES (symbolic boolean expression solver) or something. Or just
boolexpr. Whatever.