[rebase] working on desequentialization
The boolexpr struct is now the way to go to canonicalize boolean
expressions. Desequentialize now builds a list of signals and assembles
the prototype for a function capable of calculating the value and drive
condition of all sequential signals.

Now a function is needed that builds up a set of dependencies starting
from one or more values. This set shall then be used to perform a
partial copy of the original process -- only values in the set are
retained.