[rebase] working on desequentialization
Sequential signals are properly determined and the subset of
instructions required to calculate the driven value and drive condition
is calculated. A new function is instantiated that has the right input
and output parameters.

The above function now needs to be populated with the instructions that
perform the calculations. This requires some additional functions, e.g.
to copy instructions and replacing their arguments, possibly all in one
go.