Add load, store, and var instructions
Variables and memory allocation has been suspiciously absent from the
assembly until now. Add load and store instructions to read and write
from and to memory locations. Also add a var instruction that allows
variables/memory to be allocated on the stack, similar to the alloca
instruction in LLVM and the C library.