Add probe instruction
The current implementation of signals breaks the single static
assignment principle in that a signal may change while a function is
executing, e.g. during a wait instruction. To alleviate this, implement
a probe instruction that explicitly samples the state of a signal at a
specific location. Probe/drive thus gain the same semantic as load/store
in that they operate on an opaque value (signal and pointer) and return
or change the state of that value. Although this seems to bloat the code
at first glance, it should prove to be a crucial step to keep assembly
easy to optimize.