Add CLKCNT example
The assembly language still lacks a wait statement that would allow
processes to halt and wait for a specific amount of time to pass or one
or more signals to change. Add an example that contains a simple
testbench generating a clock signal and a sequential process
implementing a counter. The addition of a wait instruction is likely to
require a new kind of block that represents a piece of code during which
no time passes. Otherwise circuit analysis becomes very tricky, since
wait statements can introduce arbitrary passage of time.