add types and basic module
Add the fundamental types to `src/ty.rs`. They are packaged in an enum,
which in turn is wrapped in an arc. This allows for complicated types to
be shared, and causes type clones to have the complexity of an atomic
integer increment, rather than multiple memory allocations.
Optimizations are still possible.

Add a basic `Module` struct which encapsulates all the information
related to a design hierarchy. The module acts as a factory for types
and any root objects (entities, processes, and functions).

Also add a small test.