impl value refs, naming, block emission
Assign unique IDs to the values in the LLHD graph. Add a
`FunctionContext` which allows such an ID to be resolved to a concrete
value. The value can then be queried for its name, type, or other
information.

Change the `Visitor` to make use of the context in order to visit a
function's basic blocks. Now a similar mechanism is required for the
instructions.

Extend the assembly writer to make use of the improved visitor pattern,
now being able to print the name of the basic blocks of a function. Also
add a name uniquification mechanism to the writer, which tries to stick
as close to the given name of a value, but might resort to extending it
with a monotonically increasing number.

Make type creation a global operation. Rather than being local to a
module, there should rather be a global repository of types that is used
to ensure that for each type there is exactly one instance of it around
in memory. Everything in the graph holds an arc to such types. Currently
this mechanism is not implemented. The type creation functions simply
return a newly allocated type.