impl inst emission, refactor nodes
Extend the assembly writer to now properly emit instructions. Add a
`Context` and `UnitContext` trait that allows for value references to be
resolved when traversing the graph. Have `Visitor` make use of contexts.
Add option for values to be local or global, and adjust the generated
names accordingly. Refactor the code a bit such that arguments and
functions now live in their own source files.