impl processes and add roadmap
Implement processes. Since these shared a fair bit of semantics with
functions, add a `SeqBody` struct that implements the management of
blocks and instructions. Also implement the `Value` trait for functions
and processes. This will make it possible to use them as arguments to
instructions.

Add a roadmap to the README file which roughly outlines what needs to be
implemented in order to be able to write the first programs that make
use of LLHD.