add inst and call instructions
Add the `call` instruction to transfer control to a function and obtain
its return value. Also add the `inst` instruction to instantiate a
process or entity that will react to changes in its input signals by
driving scheduling events on its output signals. For this to work, add
a `ModuleContext` that allows functions, processes, and entities to be
resolved given their `ValueRef`.