impl assembly parsing
Add a parser that reads the human readable assembly emitted by the
corresponding writer. In its current state it is very basic, and built
around the "combine" parser combinators library. In the long term, we
might want to substitute this for a more classical lexer&parser
combination. But it will do for now.