fix empty inst arg list parsing, extend README
The parser currently breaks when an `inst` instruction has no input or
output signals. Add a separate parser that tries to parse an empty
argument list first, such that the code does not try to determine the
first argument's type, which does not exist in the empty case.

Also extend the README with some things that need to be done.