make epsilon/delta time steps usize
Change the type of the epsilon and delta time steps held within a time
constant to `usize`. This is reasonable since these values are linked to
number of iterations performed during execution, which is bound by
feasible simulation runtimes. This in contrast to the time value in
seconds, which may step in large strides.