tcm: impl drv motion
Implement moving of `drv` instructions down to the tail blocks of the
encompassing temporal region. This is pretty complex. Currently we split
the process into two phases:

1. Push down the `drv` instructions across *diverging* control flow
   until no further movement is possible. This splits drives into
   multiple instructions. Sometimes requires for a helper block to be
   introduced.
2. Merge `drv` instructions across *reconverging* control flow. This
   reduces the number of drive instructions.

The whole process is tedious and slow, but currently gets the job done.
A better approach would be to build a data structure of all drives
and branches in a DFG/CFG, then precompute the entire movement on that
data structure, and only eventually move the instructions around.