dce: improve impl and results
Improve the implementation of the Dead Code Elimination pass. Build a
separate "data structure" which tracks trivial blocks and trivial
branches, and use that to determine the entire set of branches to be
removed in the program. Then go in and replace trivial branches and
blocks. Then prune unreachable blocks that were detached by the above.
Then additionally merge blocks that trivially follow each other.